Add support for OpenSHMEM.
This implementation differs from plain Cray/SGI shmem, in that
it supports shmem_wait().
The OpenSHMEM definition specifies, that packets may be sent
in any order; therefore checking the last element for a known
value is not valid.

Instead signal completion using a flag -- and check using
shmem_wait() which of course adds to the latency of small msg.
This is default when compiling for OpenSHMEM.
The old busy-waiting behaviour may be achieved by undefining
SHMEM_WAIT.
The "friendly" busy-waiting is selected by defining SCHED_YIELD.
Otherwise busy-waiting for the value is done.
Using SCHED_DEBUG, the letters "d", "r" and "t" are output
every 100 mio. iteration.

Please note, that compared to Cray/SGI implementation, the
calls for single element transfer shmem_int_p() and
shmem_double_p() are being used.